requirements.txt
watchdog
httpx
requests
pypdf
python-docx
